python/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "sentinel-atl"
version = "0.1.2"
description = "The Agent Trust Layer — identity, credentials, and reputation for AI agents (Python SDK)"
readme = "README.md"
license = { text = "Apache-2.0" }
requires-python = ">=3.10"
authors = [
  { name = "Sentinel ATL", email = "sentinel-atl@users.noreply.github.com" },
]
keywords = [
  "ai-agent",
  "trust",
  "identity",
  "did",
  "verifiable-credentials",
  "mcp",
  "security",
  "reputation",
]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: Apache Software License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Topic :: Security",
  "Topic :: Software Development :: Libraries",
]
dependencies = [
  "pynacl>=1.5.0",
  "multiformats>=0.3.0",
  "httpx>=0.27.0",
]

[project.optional-dependencies]
langchain = ["langchain-core>=0.3.0"]
crewai = ["crewai>=0.50.0"]
openai = ["openai>=1.0.0"]
all = ["langchain-core>=0.3.0", "crewai>=0.50.0", "openai>=1.0.0"]
dev = ["pytest>=8.0", "pytest-asyncio>=0.24.0"]

[project.urls]
Homepage = "https://github.com/sentinel-atl/project-sentinel"
Documentation = "https://github.com/sentinel-atl/project-sentinel#readme"
Repository = "https://github.com/sentinel-atl/project-sentinel"
Issues = "https://github.com/sentinel-atl/project-sentinel/issues"
